// src/types/auth.ts
export type UserRole = "ADMIN" | "CONSULTANT" | "CLIENT";

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  createdAt?: string;
  updatedAt?: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface RegisterCredentials {
  name: string;
  email: string;
  password: string;
  role?: UserRole;
}

export interface BackendResponse<T> {
  statusCode: number;
  success: boolean;
  message: string;
  data: T;
}